Analyze NYSE Data
This project introduced me to the data analysis process that I will be using throughout the rest of the Nanodegree program. In this project, I went through the process of calculating summary statistics, drawing inference from the statistics, calculating business metrics and using models to forecast future growth prospects for the companies. The goal is to perform an analysis and also to create visual tools to communicate the results in informative ways.